refactor: add patch builder and drop geometry utilities
